5/2/21
Main Changes:
Change 1:
Replaced non-working search button with a cart button that takes the user to a new (non-working) cart page where the user can go to a sample checkout page.

Change 2:
Added a contact page that includes some template contact information.

Change 3:
Added 3 sample products to the "Featured Products" section on the homepage.

Change 4:
Removed unnecessary content from the concierge page and added a 20px margin to the paragraph and h2-h6.

Change 5:
Replaced the still image on the homepage with 3 new images that are in a slideshow format.

Small Changes:
-Updated nav bar links to support redirecting to new pages

-Updated certain page titles to represent what the page is. (Example: Changing the "Cart" page title to "Cart".)

-Changed nav bar "HOME" to "Home"

-Added a temporary copyright notice to the bottom of most pages. This will be removed when our own images are added.

-index.html now redirects to home.html. This may aid in getting the website on google search.